refactor(history): extract CompositeSearchInput and useCompositeSearchBar hook
Extract CompositeSearchBar's 365 lines into a composition layer and two
reusable abstractions:
- CompositeSearchInput: purely presentational component handling UI layout
  (input, chips, suggestion panel, keyboard navigation)
- useCompositeSearchBar: logic hook managing search state (buffer, highlight,
  keyboard handling, filter application, candidate building)

This separation of concerns improves testability (mock hook state in component
tests, focus logic tests independently) and reusability (hook can back
alternative UIs). Includes comprehensive test suite for CompositeSearchBar
covering keyboard navigation, filter application, and suggestion semantics.
